Closes #79 - chainstats display tolerancing
- for small numbers (~e-12) the display should correctly switch to a scientific format.
- Note, this could be improved to by letting the user control the level of accuracy, but for now the user can still access the raw data by specifying "returnstats=True" when calling the chainstats function.
